pyproject.toml
[project]
name = "pol-eth-bridge-test-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "web3>=7.8.0",
]
